Homepage - Value Proposition - Iteration 1 #13
