--- client/containers/About.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const About = () => (
-  <h1>About</h1>
-);
-
-export default About;
