--- client/containers/Home.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Home = () => (
-  <h1>Home</h1>
-);
-
-export default Home;
