--- client/routes.js
@@ -1,26 +1,0 @@
-
-/* import React from 'react';
-import {
-  BrowserRouter as Router,
-  Route,
-  Link
-} from 'react-router-dom';
-
-import Home from './containers/Home';
-import About from './containers/About';
-
-const RouterConfig = () => (
-  <Router>
-    <div>
-      <ul>
-        <li><Link to="/">Home</Link></li>
-        <li><Link to="/about">About</Link></li>
-      </ul>
-
-      <Route exact path="/" component={Home} />
-      <Route exact path="/about" component={About} />
-    </div>
-  </Router>
-);
-
-export default RouterConfig; */
